refactor: ValidateController에서 Validate 분리
